Split reverb left/right into different Ts
The hardware calculates left and right in alternating Ts at 22050Hz.
This should correct any issues with the reverb work area size needing to be
two times larger due to sampling at 41000Hz.

There can still be some aliasing if any signals above 11025Hz are added to
the wet mix. So implementing the low-pass filter from the hardware is needed.