Fix sample interpolation
Sample interpolation was subtly broken by keeping a history of interpolated
samples which is shifted back by one each tick.
This means that any voices with a very low pitch would fill the history with
the same sample and break the interpolation.
And samples with a very high pitch would jump over samples and break the
interpolation.

Instead mimic what hardware does by interpolating between the current sample
and previous samples from the current block.
This means that the last 3 samples from the previous block have to be kept
for interpolation of the first 3 samples from the current block.

This significantly improves audio quality for voices with high/low pitch,
which previously had a noisy or buzzing quality.

Also replace the divides by 64 in the ADPCM decoding with >> 6.
This gets rid of incorrect rounding if the result is negative.