Add support for drum table
A wave table entry with a lower address than the previous entry marks the
beginning of the drum table.
Originally this data is loaded to a separate array and indices >= 512 index
into it.
MGS2 takes a slightly different approach and indexes directly at a hardcoded
offset into the memory buffer as there is only one file using a drum table.
Here an index is used instead, and 512 is subtracted from the value.